+++ eslint.config.js
@@ -1,7 +1,7 @@
 import js from '@eslint/js';
 
 export default [
-  { ignores: ['dist/**', 'node_modules/**'] },
+  { ignores: ['coverage/**', 'dist/**', 'node_modules/**'] },
   js.configs.recommended,
   {
     files: ['**/*.{js,mjs}'],